requirements.txt
# App
streamlit>=1.37.0
pillow>=10.3.0
pandas>=2.2.2

# Gemini SDKs (2.x aware)
google-generativeai>=0.8.5
google-ai-generativelanguage>=0.6.6

# Observability
structlog>=24.1.0
prometheus-client>=0.20.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-otlp>=1.27.0
opentelemetry-instrumentation-requests>=0.48b0
